fix on many things

+++ src/PacketLobbyInfo.h
@@ -10,7 +10,7 @@
 struct LobbyInfoData
 {
     uint8_t m_aiControlled; // Whether the vehicle is AI (1) or Human (0) controlled
-    uint8_t m_teamId; // Team id - see appendix (255 if no team currently selected)
+    uint16_t m_teamId; // Team id - see appendix (255 if no team currently selected)
     uint8_t m_nationality; // Nationality of the driver
     uint8_t m_platform; // 1 = Steam, 3 = PlayStation, 4 = Xbox, 6 = Origin, 255 = unknown
     char m_name[cs_maxParticipantNameLen]; // Name of participant in UTF-8 format – null terminated
